pyproject.toml
[project]
name = "newsbreak-marketing"
version = "0.1.0"
description = ""
authors = [
    {name = "Saurabh",email = "sumansaurabh@ideaclan.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]

[tool.poetry]
packages = [{include = "newsbreak_marketing", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
strawberry-graphql = {extras = ["asgi"], version = "^0.275.5"}
uvicorn = "^0.35.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
